+++ spl/src/associated_token.rs
@@ -3,10 +3,12 @@ use anchor_lang::solana_program::pubkey::Pubkey;
 use anchor_lang::Result;
 use anchor_lang::{context::CpiContext, Accounts};
 
-pub use spl_associated_token_account::{get_associated_token_address, ID};
+pub use spl_associated_token_account::{
+    get_associated_token_address, instruction::create_associated_token_account, ID,
+};
 
 pub fn create<'info>(ctx: CpiContext<'_, '_, '_, 'info, Create<'info>>) -> Result<()> {
-    let ix = spl_associated_token_account::create_associated_token_account(
+    let ix = create_associated_token_account(
         ctx.accounts.payer.key,
         ctx.accounts.authority.key,
         ctx.accounts.mint.key,